+++ doc/changes.rst
@@ -20,7 +20,10 @@ New in 0.40
 
 * :func:`boosting` now accepts data with ragged trials (trials of different lengths).
 * :func:`boosting` now stores both the l1 and l2 loss of the final fit.
-* API: Plotting parameters ``ncol`` and ``nrow`` have been renamed to ``columns`` and ``rows``
+* API:
+
+   - Plotting parameters ``ncol`` and ``nrow`` have been renamed to ``columns`` and ``rows``.
+   - :class:`Factor` cells that are not specified in ``labels`` are now ordered by their first ocurrence in ``x`` (previously order was alphabetic).
 
 
 New in 0.39